requirements.txt
pymysql>=1.1.0,<2.0.0
PyYAML>=6.0,<7.0
requests>=2.31,<3.0
